clean code and make utils.h
