+++ TestShell/commandParser.cpp
@@ -1,6 +1,5 @@
 #include <iostream>
-#include <sstream>
-#include <stdexcept>
+#include "utils.h"
 #include "commandParser.h"
 
 bool CommandParser::ParseCommand(const string& fullCmd) {
@@ -25,6 +24,7 @@ bool CommandParser::ExecuteSsdUsingParsedCommand(ISsdApp* app) {
 bool CommandParser::IsVaildCommand(const string& cmd, size_t tokenSize) {
     if (cmd_set.find(cmd) != cmd_set.end()) {
         if (cmd == WRITE_CMD && tokenSize <= 2) return false;
+        if (cmd == READ_CMD && tokenSize > 2) return false;
         return true;
     }
     return false;
@@ -37,38 +37,27 @@ bool CommandParser::doParse(const string& fullCmd) {
 
     command = tokens[0];
     if (IsVaildCommand(command, tokens.size()) == false) return false;
-    if (tokens.size() > 1) lba = stringToUnsignedInt(tokens[1]);
-    if (tokens.size() > 2) data = stringToUnsignedInt(tokens[2]);
+    if (tokens.size() > 1) {
+        try {
+            lba = stringToUnsignedInt(tokens[1]);
+        }
+        catch (const std::exception& e) {
+            cout << e.what() << endl;
+            return false;
+        }
+    }
+    
+    if (tokens.size() > 2) {
+        try {
+            data = stringToUnsignedInt(tokens[2]);
+        }
+        catch (const std::exception& e) {
+            cout << e.what() << endl;
+            return false;
+        }
+    }
 
     return true;
 }
 
-std::vector<string> CommandParser::splitCommand(const string& fullCmd) {
-    std::vector<string> tokens = {};
-    std::istringstream iss(fullCmd);
-    std::string token;
-    while (iss >> token) {
-        tokens.push_back(token);
-    }
-    return tokens;
-}
 
-unsigned int CommandParser::stringToUnsignedInt(const string& str) {
-    try {
-        size_t pos = 0;
-        unsigned long value = std::stoul(str, &pos, 0);
-
-        if (pos != str.size()) {
-            throw std::invalid_argument("문자열 중간에 숫자 이외의 문자가 있음");
-        }
-        if (value > std::numeric_limits<unsigned int>::max()) {
-            throw std::out_of_range("값이 unsigned int 범위를 초과함");
-        }
-
-        return static_cast<unsigned int>(value);
-    }
-    catch (const std::exception& e) {
-        std::cerr << "변환 실패: " << e.what() << std::endl;
-        throw std::runtime_error(e.what());
-    }
-}
